Add edit room modal
