+++ src/components/RoomCard/RoomCard.jsx
@@ -1,9 +1,25 @@
 import React from 'react';
 import moment from 'moment';
 import { useNavigate } from 'react-router';
-import { EditOutlined , QrCode, FileDownloadOutlined } from '@mui/icons-material';
-import { Box, Card, CardContent, Chip, Grid, Typography, AvatarGroup } from '@mui/material';
-import { FlexBoxSpaceBetween, FlexBoxCenter, FlexBoxCenterColumn } from '../styled/general';
+import {
+  EditOutlined,
+  QrCode,
+  FileDownloadOutlined,
+} from '@mui/icons-material';
+import {
+  Box,
+  Card,
+  CardContent,
+  Chip,
+  Grid,
+  Typography,
+  AvatarGroup,
+} from '@mui/material';
+import {
+  FlexBoxSpaceBetween,
+  FlexBoxCenter,
+  FlexBoxCenterColumn,
+} from '../styled/general';
 import { REFLECTION_ID_MAP } from '../../models/storyMap';
 import { StyledRoomCardTag } from '../RoomCardTag/StyledRoomCardTag';
 import { palette } from '@mui/system';
@@ -11,7 +27,7 @@ import { CharacterAvatar } from '../CharacterAvatar/CharacterAvatar';
 import { CharacterAvatarGroup } from '../CharacterAvatarGroup/CharacterAvatarGroup';
 
 const RoomCard = (props) => {
-  const { room, handleDelete, toggleIsActive } = props;
+  const { room, handleDelete, handleEdit, toggleIsActive } = props;
   const navigate = useNavigate();
   const firstReflectionId = room.reflectionIds[0];
 
@@ -30,42 +46,74 @@ const RoomCard = (props) => {
         navigate(`/room/${room.id}/reflectionId/${firstReflectionId}`)
       }
     >
-        <Grid container sx={{padding: "36px 40px", height: "100%" }} spacing={1}>
-          <FlexBoxSpaceBetween sx={{width: "100%", height: "28px"}}>
-            <StyledRoomCardTag
-              label={'TODO: upcoming/archived'}
-              variant='outlined'
-              size='small'
+      <Grid container sx={{ padding: '36px 40px', height: '100%' }} spacing={1}>
+        <FlexBoxSpaceBetween sx={{ width: '100%', height: '28px' }}>
+          <StyledRoomCardTag
+            label={'TODO: upcoming/archived'}
+            variant='outlined'
+            size='small'
+          />
+          <FlexBoxCenter>
+            <EditOutlined
+              sx={{
+                color: (theme) => theme.palette.lapis[100],
+                marginLeft: '4px',
+              }}
+              onClick={(e) => {
+                e.stopPropagation();
+                handleEdit();
+              }}
             />
-            <FlexBoxCenter>
-              <EditOutlined sx={{color: (theme) => theme.palette.lapis[100], marginLeft: "4px"}}/>
-              <QrCode sx={{color: (theme) => theme.palette.lapis[100], marginLeft: "4px"}}/>
-              <FileDownloadOutlined sx={{color: (theme) => theme.palette.lapis[100], marginLeft: "4px"}}/>
-            </FlexBoxCenter>
-          </FlexBoxSpaceBetween>
-          <Grid item xs={12}>
-            <Typography variant="h6" sx={{ textTransform: 'uppercase', opacity: 0.5, fontSize: 14 }}>
-              {dateString}
+            <QrCode
+              sx={{
+                color: (theme) => theme.palette.lapis[100],
+                marginLeft: '4px',
+              }}
+            />
+            <FileDownloadOutlined
+              sx={{
+                color: (theme) => theme.palette.lapis[100],
+                marginLeft: '4px',
+              }}
+            />
+          </FlexBoxCenter>
+        </FlexBoxSpaceBetween>
+        <Grid item xs={12}>
+          <Typography
+            variant='h6'
+            sx={{ textTransform: 'uppercase', opacity: 0.5, fontSize: 14 }}
+          >
+            {dateString}
+          </Typography>
+        </Grid>
+        <Grid
+          item
+          xs={7}
+          sx={{ paddingRight: '8px', height: '100%' }}
+          direction='column'
+          display='flex'
+          justifyContent='space-between'
+        >
+          <Grid item>
+            <Typography variant='h4' sx={{ mb: '8px' }}>
+              {room.name}
+            </Typography>
+            <Typography>Class Code: {room.code}</Typography>
+          </Grid>
+          <Grid>
+            <Typography
+              variant='body2'
+              sx={{ textTransform: 'uppercase', marginBottom: '52px' }}
+            >
+              {room.organisation}
             </Typography>
           </Grid>
-          <Grid item xs={7} sx={{paddingRight: "8px", height: "100%"}} direction="column" display="flex" justifyContent="space-between">
-              <Grid item>
-                <Typography variant='h4' sx={{ mb: "8px"}}>{room.name}</Typography>
-                <Typography>Class Code: {room.code}</Typography>
+        </Grid>
 
-              </Grid>
-              <Grid>
-                <Typography variant='body2' sx={{ textTransform: 'uppercase', marginBottom: "52px" }}>
-                  {room.organisation}
-                </Typography>
-              </Grid>
-          </Grid>
-          
-          {/* right side */}
-          <Grid item xs={5}>
-
-            <CharacterAvatarGroup data={room.reflectionIds}/>
-            {/* {room.reflectionIds.map((reflectionId) => {
+        {/* right side */}
+        <Grid item xs={5}>
+          <CharacterAvatarGroup data={room.reflectionIds} />
+          {/* {room.reflectionIds.map((reflectionId) => {
               const { character, chapter } = REFLECTION_ID_MAP[reflectionId];
               return (
                 <Typography key={reflectionId}>
@@ -73,30 +121,30 @@ const RoomCard = (props) => {
                 </Typography>
               );
             })} */}
-            <Typography>
-              <span
-                style={{ cursor: 'pointer' }}
-                onClick={(e) => {
-                  e.stopPropagation();
-                  toggleIsActive();
-                }}
-              >
-                Toggle
-              </span>
-            </Typography>
-            <Typography>
-              <span
-                style={{ cursor: 'pointer' }}
-                onClick={(e) => {
-                  e.stopPropagation();
-                  handleDelete();
-                }}
-              >
-                Delete
-              </span>
-            </Typography>
-          </Grid>
+          <Typography>
+            <span
+              style={{ cursor: 'pointer' }}
+              onClick={(e) => {
+                e.stopPropagation();
+                toggleIsActive();
+              }}
+            >
+              Toggle
+            </span>
+          </Typography>
+          <Typography>
+            <span
+              style={{ cursor: 'pointer' }}
+              onClick={(e) => {
+                e.stopPropagation();
+                handleDelete();
+              }}
+            >
+              Delete
+            </span>
+          </Typography>
         </Grid>
+      </Grid>
     </Card>
   );
 };
