+++ src/components/styled/Dashboard/generalRoomModal.jsx
