+++ src/page/Home.jsx
@@ -1,31 +1,46 @@
-import React, { useCallback, useEffect, useState } from "react";
-import moment from "moment";
-import { ToggleButton, Typography, Box, Grid, Paper, MenuItem } from "@mui/material";
-import { useAuth } from "../contexts/AuthContext";
-import { deleteDbRoom, getDbRooms, updateDbRoom } from "../models/roomModel";
-import RoomCard from "../components/RoomCard/RoomCard";
-import NewRoomModal from "../components/dashboard/NewRoomModal";
+import React, { useEffect, useState } from 'react';
+import moment from 'moment';
+import { ToggleButton, Typography, Box, Grid, Paper } from '@mui/material';
+import { useAuth } from '../contexts/AuthContext';
+import { deleteDbRoom, getDbRooms, updateDbRoom } from '../models/roomModel';
+import { getDbUser } from '../models/userModel';
+import RoomCard from '../components/RoomCard/RoomCard';
+import NewRoomModal from '../components/dashboard/NewRoomModal';
+import EditRoomModal from '../components/dashboard/EditRoomModal';
 import {
   FlexBoxSpaceBetween,
-  FlexBoxCenter
-} from "../components/styled/general";
-import { GeneralButton } from "../components/GeneralButton/GeneralButton";
-import { HomeToggleButtonGroup } from "../components/HomeToggleButtonGroup/HomeToggleButtonGroup";
-import { GeneralSelect } from "../components/GeneralSelect/GeneralSelect";
+  FlexBoxCenter,
+} from '../components/styled/general';
+import { GeneralButton } from '../components/GeneralButton/GeneralButton';
+import { HomeToggleButtonGroup } from '../components/HomeToggleButtonGroup/HomeToggleButtonGroup';
+
+const formDataTemplate = {
+  name: '',
+  organisation: '',
+  reflectionIds: [],
+  date: moment().format('YYYY-MM-DD'),
+  instructions: '',
+};
 
 const Home = () => {
   const { currentUser } = useAuth();
   const [rooms, setRooms] = useState(null);
   const [filters, setFilters] = useState(() => ['all']);
   const [isNewRoomModalOpen, setIsNewRoomModalOpen] = useState(false);
+  const [isEditRoomModalOpen, setIsEditRoomModalOpen] = useState(false);
+  const [newRoomFormData, setNewRoomFormData] = useState(formDataTemplate);
+  const [editRoomFormData, setEditRoomFormData] = useState(formDataTemplate);
+  // Used to reset the new room modal's form data after a room is created.
+  // Its organisation field will be updated to match user.organisation once it's loaded.
+  const [initialRoomFormData, setInitialRoomFormData] =
+    useState(formDataTemplate);
 
   // const [roomFilter, setRoomFilter] = useState({none: "Filter"})
-  // const filterOptions = 
-  //   { 
-  //     all: "All", 
+  // const filterOptions =
+  //   {
+  //     all: "All",
   //     upcoming: "Upcoming"
   //   }
-  
 
   // const handleChange = (event) => {
   //   setRoomFilter({ [event.target.value]: filterOptions[event.target.value] });
@@ -34,6 +49,25 @@ const Home = () => {
   async function handleDeleteRoom(id) {
     await deleteDbRoom(id);
     setRooms(rooms.filter((room) => room.id !== id));
+  }
+
+  async function handleNewRoom() {
+    setNewRoomFormData(initialRoomFormData); // This is optional: wipes existing data when we re-open the new room modal
+    setIsNewRoomModalOpen(true);
+  }
+
+  async function handleEditRoom(id) {
+    const room = rooms.find((room) => room.id === id);
+    setEditRoomFormData({
+      name: room.name,
+      organisation: room.organisation,
+      reflectionIds: room.reflectionIds,
+      date: moment(room.date).format('YYYY-MM-DD'),
+      instructions: room.instructions,
+      // Also pass in ID into formData, necessary for editing the room
+      id: room.id,
+    });
+    setIsEditRoomModalOpen(true);
   }
 
   async function toggleIsActiveRoom(id) {
@@ -49,20 +83,31 @@ const Home = () => {
     setFilters(newFilters);
   }
 
-  const loadRooms = useCallback(async () => {
+  const loadRooms = async () => {
     const facilitatorId = currentUser.id;
     const rooms = await getDbRooms(facilitatorId);
     setRooms(rooms);
-  }, [currentUser]);
+  };
+
+  const loadUserOrganisation = async () => {
+    const dbUser = await getDbUser(currentUser.id);
+    // Set the new room's organisation to be user's organisation by default
+    setNewRoomFormData({
+      ...newRoomFormData,
+      organisation: dbUser.organisation,
+    });
+    setInitialRoomFormData({
+      ...initialRoomFormData,
+      organisation: dbUser.organisation,
+    });
+  };
 
   useEffect(loadRooms, []);
+  useEffect(loadUserOrganisation, []);
 
   const AddClassButton = () => {
     return (
-      <GeneralButton
-        variant='special'
-        onClick={() => setIsNewRoomModalOpen(true)}
-      >
+      <GeneralButton variant='special' onClick={handleNewRoom}>
         Add a class
       </GeneralButton>
     );
@@ -74,7 +119,7 @@ const Home = () => {
     } else if (rooms.length === 0) {
       return (
         <Box>
-          <Typography variant="h5">No classes yet</Typography>
+          <Typography variant='h5'>No classes yet</Typography>
           <Typography>Get started by starting a class!</Typography>
           <AddClassButton />
         </Box>
@@ -90,6 +135,7 @@ const Home = () => {
         }
         return true;
       });
+      // TODO: sort rooms by some order
       return (
         <Grid container spacing={2}>
           {filteredRooms.map((room) => (
@@ -98,6 +144,7 @@ const Home = () => {
                 room={room}
                 toggleIsActive={() => toggleIsActiveRoom(room.id)}
                 handleDelete={() => handleDeleteRoom(room.id)}
+                handleEdit={() => handleEditRoom(room.id)}
               />
             </Grid>
           ))}
@@ -107,12 +154,14 @@ const Home = () => {
   };
 
   return (
-    <Box sx={{height: "100%"}}>
-      <Paper sx={{position: "absolute", left: 0, right: 0, padding: "20px 68px"}}>
+    <Box sx={{ height: '100%' }}>
+      <Paper
+        sx={{ position: 'absolute', left: 0, right: 0, padding: '20px 68px' }}
+      >
         <FlexBoxSpaceBetween>
           <Box>
             <Typography variant='h4' sx={{ marginBottom: '12px' }}>
-            Your classes
+              Your classes
             </Typography>
             <FlexBoxCenter>
               <HomeToggleButtonGroup
@@ -135,7 +184,7 @@ const Home = () => {
                   ARCHIVED
                 </ToggleButton>
               </HomeToggleButtonGroup>
-                {/* <GeneralSelect
+              {/* <GeneralSelect
                   name="filter"
                   value={roomFilter.key}
                   onChange={handleChange}
@@ -153,17 +202,34 @@ const Home = () => {
           <AddClassButton />
         </FlexBoxSpaceBetween>
       </Paper>
-      <Box sx={{ padding: '68px', paddingTop: "140px", background: (theme) => theme.palette.lapis[10], height: "calc(100vh - 120px)" }}>
+      <Box
+        sx={{
+          padding: '68px',
+          paddingTop: '140px',
+          background: (theme) => theme.palette.lapis[10],
+          height: 'calc(100vh - 120px)',
+        }}
+      >
         <Box>
           <Rooms />
         </Box>
         <NewRoomModal
-          isNewRoomModalOpen={isNewRoomModalOpen}
-          setIsNewRoomModalOpen={setIsNewRoomModalOpen}
+          isModalOpen={isNewRoomModalOpen}
+          setIsModalOpen={setIsNewRoomModalOpen}
+          formData={newRoomFormData}
+          setFormData={setNewRoomFormData}
+          initialFormData={initialRoomFormData}
+          loadRooms={loadRooms}
+        />
+        <EditRoomModal
+          isModalOpen={isEditRoomModalOpen}
+          setIsModalOpen={setIsEditRoomModalOpen}
+          formData={editRoomFormData}
+          setFormData={setEditRoomFormData}
+          initialFormData={initialRoomFormData}
           loadRooms={loadRooms}
         />
       </Box>
-
     </Box>
   );
 };
